requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
Django==5.1
fake-useragent==1.5.1
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
kaitaistruct==0.10
kombu==5.4.0
msgspec==0.18.6
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
pyasn1==0.6.0
pycparser==2.22
pydivert==2.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-rucaptcha==6.2
requests==2.32.3
selenium==4.23.1
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
StrEnum==0.4.15
tenacity==8.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
zstandard==0.23.0